refactor: added setup function + moved all functions to api module
Now the user can at any time configure the plugin via the setup function.
The three commands have been moved to an API module, we could extend them to be also callable by `lua` later, via `require("gist.api.create").from_buffer` and so on.
The README has been updated, I can't test `packer` package manager but it should be working (lazy) is.
