+++ plugin/gist.lua
@@ -1,12 +1,12 @@
-local gist = require("gist")
+local gist = require("gist.api")
 
-vim.api.nvim_create_user_command("CreateGistFromFile", gist.create_from_file, {
+vim.api.nvim_create_user_command("CreateGistFromFile", gist.create_from_buffer, {
 	nargs = "?",
 	desc = "Create a Gist from the current file.",
 	range = false,
 })
 
-vim.api.nvim_create_user_command("CreateGist", gist.create, {
+vim.api.nvim_create_user_command("CreateGist", gist.create_from_file, {
 	nargs = "?",
 	desc = "Create a Gist from the current selection.",
 	range = true,
